#9492: move matmul code to ttnn directory hierarchy
