+++ tt_eager/tt_dnn/op_library/complex/complex_ops.cpp
@@ -4,7 +4,6 @@
 
 #include "tt_dnn/op_library/complex/complex_ops.hpp"
 #include "tt_dnn/op_library/concat/concat_op.hpp"
-#include "tt_dnn/op_library/bmm/bmm_op.hpp"
 #include "tt_dnn/op_library/reshape/reshape_op.hpp"
 #include "ttnn/operations/data_movement/slice/slice.hpp"
 #include "tt_numpy/functions.hpp"
