+++ ttnn/cpp/ttnn/operations/matmul/device/matmul_op.cpp
@@ -2,7 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include "tt_dnn/op_library/bmm/bmm_op.hpp"
+#include "ttnn/operations/matmul/device/matmul_op.hpp"
 
 #include <algorithm>
 #include <cmath>
