+++ ttnn/cpp/ttnn/operations/matmul/device/multi_core_reuse/bmm_op_multi_core_reuse.cpp
@@ -2,24 +2,19 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include "tt_dnn/op_library/bmm/bmm_op.hpp"
+#include "ttnn/operations/matmul/device/matmul_op.hpp"
+
 #include "tt_dnn/op_library/work_split.hpp"
 #include "tt_dnn/op_library/operation.hpp"
-
 #include "tt_metal/host_api.hpp"
 #include "tt_metal/common/constants.hpp"
 #include "tt_metal/detail/util.hpp"
 
 using namespace tt::constants;
 using namespace tt;
+using tt_metal::Buffer;
 
-
-namespace reuse_padding_helpers {
-using namespace tt::constants;
-using namespace tt;
-using namespace tt_metal;
-
-operation::ProgramWithCallbacks create_program(
+tt_metal::operation::ProgramWithCallbacks create_program(
     tt_metal::Device *device,
     tt::DataFormat cb_data_format,
     MathFidelity math_fidelity,
@@ -76,42 +71,33 @@ operation::ProgramWithCallbacks create_program(
         B // batch
     };
 
-    // Parameters for last row, col, or block
-    uint32_t last_block_h = M % per_core_M == 0 ? per_core_M : M % per_core_M;
-    uint32_t last_block_w = N % per_core_N == 0 ? per_core_N : N % per_core_N;
-    uint32_t last_block_num_nonzero_subblocks_h = (last_block_h  - 1) / out_subblock_h + 1;
-    uint32_t last_block_num_nonzero_subblocks_w = (last_block_w  - 1) / out_subblock_w + 1;
-    uint32_t last_subblock_of_last_block_h = last_block_h % out_subblock_h == 0 ? out_subblock_h : last_block_h % out_subblock_h;
-    uint32_t last_subblock_of_last_block_w = last_block_w % out_subblock_w == 0 ? out_subblock_w : last_block_w % out_subblock_w;
-    uint32_t last_block_padded_subblock_tiles_addr_skip = single_tile_size * (out_subblock_w - last_subblock_of_last_block_w);
-    uint32_t last_block_padded_block_tiles_w_skip =  (out_subblock_w * out_subblock_h) * (per_core_N / out_subblock_w - last_block_num_nonzero_subblocks_w);
-    uint32_t last_block_padded_block_tiles_h_skip = (per_core_M / out_subblock_h - last_block_num_nonzero_subblocks_h) * (per_core_N * out_subblock_h);
-
-    uint32_t num_blocks_read = 0; // Shouldn't be used since this op is only used for 1 core
-    uint32_t num_blocks_y = (M - 1) / per_core_M + 1; // Should always be 1
-    uint32_t num_blocks_x = (N - 1) / per_core_N + 1; // SHould always be 1
+    uint32_t num_blocks_read = 0;
+    uint32_t num_blocks_y = M / per_core_M;
+    uint32_t num_blocks_x = N / per_core_N;
 
     CoreRangeSet all_cores(tt::tt_metal::num_cores_to_corerange_set(num_blocks_x * num_blocks_y, device->compute_with_storage_grid_size(), true));
+
+    // Create circular buffers
     uint32_t src0_cb_index = 0;
-    tt_metal::CircularBufferConfig cb_src0_config = tt_metal::CircularBufferConfig(in0_CB_size, {{src0_cb_index, cb_data_format}})
+    tt_metal::CircularBufferConfig src0_cb_config = tt_metal::CircularBufferConfig(in0_CB_size, {{src0_cb_index, cb_data_format}})
 		.set_page_size(src0_cb_index, single_tile_size);
-    auto cb_src0 = tt_metal::CreateCircularBuffer(program, all_cores, cb_src0_config);
+	auto cb_src0 = tt_metal::CreateCircularBuffer(program, all_cores, src0_cb_config);
 
-    uint32_t src1_cb_index = 1;
-    tt_metal::CircularBufferConfig cb_src1_config = tt_metal::CircularBufferConfig(in1_CB_size, {{src1_cb_index, cb_data_format}})
+	uint32_t src1_cb_index = 1;
+	tt_metal::CircularBufferConfig src1_cb_config = tt_metal::CircularBufferConfig(in1_CB_size, {{src1_cb_index, cb_data_format}})
 		.set_page_size(src1_cb_index, single_tile_size);
-    auto cb_src1 = tt_metal::CreateCircularBuffer(program, all_cores, cb_src1_config);
+	auto cb_src1 = tt_metal::CreateCircularBuffer(program, all_cores, src1_cb_config);
 
-    uint32_t output_cb_index = 16; // output operands start at index 16
+	uint32_t output_cb_index = 16; // output operands start at index 16
     uint32_t interm0_cb_index = 24;
-    std::map<uint8_t, tt::DataFormat> output_cb_data_format_spec = {
+    std::map<uint8_t, tt::DataFormat> output_cb_data_format_spec {
         {output_cb_index, cb_data_format},
         {interm0_cb_index, cb_data_format}
     };
-    tt_metal::CircularBufferConfig cb_output_config = tt_metal::CircularBufferConfig(out_CB_size, output_cb_data_format_spec)
+	tt_metal::CircularBufferConfig output_cb_config = tt_metal::CircularBufferConfig(out_CB_size, output_cb_data_format_spec)
 		.set_page_size(output_cb_index, single_tile_size)
         .set_page_size(interm0_cb_index, single_tile_size);
-    auto cb_output = tt_metal::CreateCircularBuffer(program, CoreRangeSet({all_cores}), cb_output_config);
+	auto cb_output = tt_metal::CreateCircularBuffer(program, all_cores, output_cb_config);
 
     bool in0_is_dram = in0_buffer->buffer_type() == tt_metal::BufferType::DRAM ? 1 : 0;
     bool in1_is_dram = in1_buffer->buffer_type() == tt_metal::BufferType::DRAM ? 1 : 0;
@@ -123,20 +109,20 @@ operation::ProgramWithCallbacks create_program(
     // Create reader and writer kernels per core
     auto mm_reader_kernel_id = tt_metal::CreateKernel(
         program,
-        "tt_eager/tt_dnn/op_library/bmm/kernels/dataflow/reader_bmm_tile_layout_padding.cpp_padding.cpp",
+        "ttnn/cpp/ttnn/operations/matmul/device/kernels/dataflow/reader_bmm_tile_layout.cpp",
         all_cores,
         tt_metal::ReaderDataMovementConfig(reader_compile_time_args));
 
     auto unary_writer_kernel_id = tt_metal::CreateKernel(
         program,
-        "tt_eager/tt_dnn/op_library/bmm/kernels/dataflow/writer_bmm_tile_layout_padding.cpp",
+        "ttnn/cpp/ttnn/operations/matmul/device/kernels/dataflow/writer_bmm_tile_layout.cpp",
         all_cores,
         tt_metal::WriterDataMovementConfig(writer_compile_time_args));
 
     // Create compute kernel
     auto mm_kernel_id = tt_metal::CreateKernel(
         program,
-        "tt_eager/tt_dnn/op_library/bmm/kernels/compute/bmm_large_block_zm.cpp",
+        "ttnn/cpp/ttnn/operations/matmul/device/kernels/compute/bmm_large_block_zm.cpp",
         all_cores,
         tt_metal::ComputeConfig{.math_fidelity = math_fidelity, .compile_args = compute_kernel_args}
     );
@@ -174,10 +160,7 @@ operation::ProgramWithCallbacks create_program(
                 (std::uint32_t)  M * K, // MtKt
                 (std::uint32_t)  K * N, // KtNt
                 (std::uint32_t)  B, // batch
-                (std::uint32_t)  bcast_batch, // bcast_B
-
-                (std::uint32_t)  last_block_h,
-                (std::uint32_t)  last_block_w
+                (std::uint32_t)  bcast_batch // bcast_B
             };
 
             std::vector<uint32_t> writer_args = {
@@ -195,15 +178,7 @@ operation::ProgramWithCallbacks create_program(
                 (std::uint32_t) (per_core_M / out_subblock_h), // out_num_subblocks_h
 
                 (std::uint32_t) M * N, // MtNt
-                (std::uint32_t) B, // batch
-
-                (std::uint32_t) last_block_num_nonzero_subblocks_h,
-                (std::uint32_t) last_subblock_of_last_block_h,
-                (std::uint32_t) last_block_padded_block_tiles_h_skip,
-                (std::uint32_t) last_block_num_nonzero_subblocks_w,
-                (std::uint32_t) last_subblock_of_last_block_w,
-                (std::uint32_t) last_block_padded_subblock_tiles_addr_skip,
-                (std::uint32_t) last_block_padded_block_tiles_w_skip
+                (std::uint32_t) B // batch
             };
 
             tt_metal::SetRuntimeArgs(program, mm_reader_kernel_id, core, mm_reader_args);
@@ -221,7 +196,7 @@ operation::ProgramWithCallbacks create_program(
         num_blocks_x
     ]
     (
-        const Program &program,
+        const tt_metal::Program &program,
         const std::vector<Buffer*>& input_buffers,
         const std::vector<Buffer*>& output_buffers
     ) {
@@ -256,20 +231,20 @@ operation::ProgramWithCallbacks create_program(
     return {std::move(program), override_runtime_args_callback};
 }
 
-}
 
 namespace tt {
 
 namespace tt_metal {
 
 
-operation::ProgramWithCallbacks matmul_multi_core_reuse_padding(const Tensor &a, const Tensor &b, Tensor& output, bool bcast_batch) {
+operation::ProgramWithCallbacks matmul_multi_core_reuse(const Tensor &a, const Tensor &b, Tensor& output, bool bcast_batch) {
 
     const auto& ashape = a.get_legacy_shape(), bshape = b.get_legacy_shape();
 
     tt::DataFormat cb_data_format = tt_metal::datatype_to_dataformat_converter(a.get_dtype());
     uint32_t single_tile_size = tt_metal::detail::TileSize(cb_data_format);
     MathFidelity math_fidelity = MathFidelity::HiFi4;
+
     tt_metal::Buffer *in0_buffer = a.buffer();
     tt_metal::Buffer *in1_buffer = b.buffer();
 
@@ -288,6 +263,8 @@ operation::ProgramWithCallbacks matmul_multi_core_reuse_padding(const Tensor &a,
     uint32_t per_core_M = 16;
     uint32_t per_core_N = 16;
 
+    TT_FATAL(Mt % per_core_M == 0);
+    TT_FATAL(Nt % per_core_N == 0);
     TT_FATAL(Kt % in0_block_w == 0);
 
     // This should allocate a DRAM buffer on the device
@@ -296,10 +273,7 @@ operation::ProgramWithCallbacks matmul_multi_core_reuse_padding(const Tensor &a,
     uint32_t num_cores_x = compute_with_storage_grid_size.x;
     uint32_t num_cores_y = compute_with_storage_grid_size.y;
 
-    // Calculate number of blocks along x and y; tensor dims are padded up to 512
-    uint32_t num_blocks_y = (Mt - 1) / per_core_M + 1; // Should always be 1
-    uint32_t num_blocks_x = (Nt - 1) / per_core_N + 1; // Should always be 1
-    uint32_t num_blocks_total = num_blocks_y * num_blocks_x;
+    uint32_t num_blocks_total = (Mt / per_core_M) * (Nt / per_core_N);
     TT_FATAL(num_blocks_total <= num_cores_x * num_cores_y);
 
     ////////////////////////////////////////////////////////////////////////////
@@ -312,7 +286,8 @@ operation::ProgramWithCallbacks matmul_multi_core_reuse_padding(const Tensor &a,
     ////////////////////////////////////////////////////////////////////////////
     //                      Application Setup
     ////////////////////////////////////////////////////////////////////////////
-    return reuse_padding_helpers::create_program(
+
+    return create_program(
         device,
         cb_data_format,
         math_fidelity,
@@ -326,7 +301,6 @@ operation::ProgramWithCallbacks matmul_multi_core_reuse_padding(const Tensor &a,
         in0_buffer, in1_buffer, out_buffer
     );
 }
-
 
 }  // namespace tt_metal
